Use long istead of double

+++ src/main/groovy/edu/oregonstate/mist/core/PersonObject.groovy
@@ -8,7 +8,7 @@ class PersonObject {
     String lastName
     String homePhone
     String alternatePhone
-    Double osuUID
+    Long osuUID
     Date birthDate
     String primaryAffiliation
     String firstName
